[Rename] 파일명 및 폴더명 변경 - pages/sotre-page -> store-page - storeitem.tsx -> index.tsx
